chore(snownet): add back span in `handle_timeout` (#10025)
Whilst entering and leaving a span for every packet is very expensive,
doing the same whenever we make timeout related changes is just fine.
Thus, we re-introduce a span removed in #9949 but only for the
`handle_timeout` function.

This gives us the context of the connection ID for not just our own
logs, but also the ones from `boringtun`.