fix(relay): swap MACs for relayed traffic
In nearly all environments, we can safely assume that we will always
use the same network gateway for forwarding relayed packets that we
received them from.

By leveraging this assumption, we can simply swap the SRC and DST MAC
addresses, removing the need to keep a HaspMap for these.

Related: #10138